Object(s) as symbolics

These functions take various R objects and return the equivalent symbols orcalls (e.g. "func" to func, "pkg::func(x, ...)" to pkg::func(x, ...), etc.`).

as_symbolic_obj() returns a symbolic.

as_symbolic_objs() returns a list of symbolics if length of input greater than one. A symbolic otherwise.

as_symbolic_obj(x, with_ns = TRUE)

as_symbolic_objs(
  ...,
  .x = NULL,
  .named = FALSE,
  .name_all = FALSE,
  .names_type = c("full", "sym", "head")
)

Arguments

x

an R object. Precisely, a syntactic literal, a symbol, a call, a function, or a formula.

...

R objects as above.

.x

a list of R objects. It gets appended to ....

.named

TRUE or FALSE. If TRUE (default), the unnamed elements of the output of as_symbolic_objs are automatically named.

.name_all

TRUE or FALSE. If TRUE (default), all elements of the output of as_symbolic_objs are automatically named.

.names_type

one of c("full", "sym", "head").

"full"

The output is equivalent to deparse(x).

"sym"

The function returns the string equivalent of a symbol x is a symbol or a call with no arguments.

"head"

The function returns only the string equivalent of a symbol or the call name.

Value

a symbolic object. That is, a symbol or a call.

Examples

as_symbolic_obj(quote(mean))
#> mean
as_symbolic_obj("mean")
#> mean
as_symbolic_obj("base::mean(1:10)")
#> base::mean(1:10)
as_symbolic_obj(quote(sd(1:10)))
#> sd(1:10)
as_symbolic_obj(quote(base::mean))
#> base::mean

as_symbolic_objs(quote(mean), "mean", "base::mean(1:10)", quote(sd(1:10)))
#> [[1]]
#> mean
#> 
#> [[2]]
#> mean
#> 
#> [[3]]
#> base::mean(1:10)
#> 
#> [[4]]
#> sd(1:10)
#> 

x <- list(quote(mean), b = "mean", c = "base::mean(1:10)", quote(sd(1:10)))
as_symbolic_objs("fn", .x = x)
#> [[1]]
#> fn
#> 
#> [[2]]
#> mean
#> 
#> $b
#> mean
#> 
#> $c
#> base::mean(1:10)
#> 
#> [[5]]
#> sd(1:10)
#> 
as_symbolic_objs("fn", .x = x, .named = TRUE)
#> $fn
#> fn
#> 
#> $mean
#> mean
#> 
#> $b
#> mean
#> 
#> $c
#> base::mean(1:10)
#> 
#> $`sd(1:10)`
#> sd(1:10)
#> 
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE)
#> $fn
#> fn
#> 
#> $mean
#> mean
#> 
#> $mean
#> mean
#> 
#> $`base::mean(1:10)`
#> base::mean(1:10)
#> 
#> $`sd(1:10)`
#> sd(1:10)
#> 
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "full")
#> $fn
#> fn
#> 
#> $mean
#> mean
#> 
#> $mean
#> mean
#> 
#> $`base::mean(1:10)`
#> base::mean(1:10)
#> 
#> $`sd(1:10)`
#> sd(1:10)
#> 
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "sym")
#> $fn
#> fn
#> 
#> $mean
#> mean
#> 
#> $mean
#> mean
#> 
#> $`base::mean(1:10)`
#> base::mean(1:10)
#> 
#> $`sd(1:10)`
#> sd(1:10)
#> 
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "head")
#> $fn
#> fn
#> 
#> $mean
#> mean
#> 
#> $mean
#> mean
#> 
#> $`base::mean`
#> base::mean(1:10)
#> 
#> $sd
#> sd(1:10)
#>