Extensions of rlang::is_symbol and rlang::is_call

Extensions of rlang::is_symbol and rlang::is_call

is_symbol2(x, name = NULL, ns_sym_is_sym = TRUE)

is_call2(x, name = NULL, n = NULL, ns = NULL, ns_sym_is_call = FALSE)

Arguments

x

An object to test. Formulas and quosures are treated literally.

name

An optional name that the call should match. It is passed to sym() before matching. This argument is vectorised and you can supply a vector of names to match. In this case, is_call() returns TRUE if at least one name matches.

ns_sym_is_sym

TRUE or FALSE. If TRUE (default), the function returns TRUE for symbols and a call of the type pkg::fun. FALSE otherwise.

n

An optional number of arguments that the call should match.

ns

The namespace of the call. If NULL, the namespace doesn't participate in the pattern-matching. If an empty string "" and x is a namespaced call, is_call() returns FALSE. If any other string, is_call() checks that x is namespaced within ns.

Can be a character vector of namespaces, in which case the call has to match at least one of them, otherwise is_call() returns FALSE.

ns_sym_is_call

TRUE or FALSE. If TRUE (default), the function returns TRUE for a calls and a call of the type pkg::fun(). FALSE otherwise.

Value

TRUE or FALSE.